+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/bean/entity/Poab.java
@@ -12,6 +12,7 @@ import com.dwarfeng.subgrade.stack.bean.entity.Entity;
 public class Poab implements Entity<PoabKey> {
 
     public static final int PERMISSION_LEVEL_OWNER = 0;
+    public static final int PERMISSION_LEVEL_GUEST = 1;
 
     private static final long serialVersionUID = 6405468025576702046L;