+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/exception/UserNotExistsException.java
@@ -1,6 +1,7 @@
 package com.dwarfeng.familyhelper.finance.stack.exception;
 
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
 
 /**
  * 用户不存在异常。
@@ -8,9 +9,9 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
  * @author DwArFeng
  * @since 1.1.0
  */
-public class UserNotExistsException extends BalanceOperateException {
+public class UserNotExistsException extends HandlerException {
 
-    private static final long serialVersionUID = -5066454985579730219L;
+    private static final long serialVersionUID = -6249161432514690530L;
 
     private final StringIdKey userKey;