+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/exception/UserNotPermittedException.java
@@ -2,6 +2,7 @@ package com.dwarfeng.familyhelper.finance.stack.exception;
 
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
 
 /**
  * 用户对账本没有权限异常。
@@ -9,9 +10,9 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
  * @author DwArFeng
  * @since 1.1.0
  */
-public class UserNotPermittedException extends BalanceOperateException {
+public class UserNotPermittedException extends HandlerException {
 
-    private static final long serialVersionUID = -5066454985579730219L;
+    private static final long serialVersionUID = -148148909478561243L;
 
     private final StringIdKey userKey;
     private final LongIdKey accountBookKey;