+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/service/BankCardOperateService.java
@@ -1,0 +1,37 @@
+package com.dwarfeng.familyhelper.finance.stack.service;
+
+import com.dwarfeng.familyhelper.finance.stack.bean.dto.BankCardCreateInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import com.dwarfeng.subgrade.stack.handler.Handler;
+
+/**
+ * 银行卡操作服务。
+ *
+ * @author DwArFeng
+ * @since 1.1.0
+ */
+public interface BankCardOperateService extends Handler {
+
+    /**
+     * 创建银行卡。
+     *
+     * @param userKey            银行卡的所有者的主键。
+     * @param accountBookKey     银行卡所属的账户的主键。
+     * @param bankCardCreateInfo 银行卡的创建信息。
+     * @return 生成的银行卡的主键。
+     * @throws ServiceException 服务异常。
+     */
+    LongIdKey createBankCard(StringIdKey userKey, LongIdKey accountBookKey, BankCardCreateInfo bankCardCreateInfo)
+            throws ServiceException;
+
+    /**
+     * 移除银行卡。
+     *
+     * @param userKey     银行卡的所有者的主键。
+     * @param bankCardKey 银行卡主键。
+     * @throws ServiceException 服务异常。
+     */
+    void removeBankCard(StringIdKey userKey, LongIdKey bankCardKey) throws ServiceException;
+}